backend-deploy/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
openai==1.3.0
pandas==1.5.3
numpy==1.24.3
joblib==1.3.2
pydantic==2.5.0
gunicorn==21.2.0
scikit-learn==1.3.0

